novel-automation/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
langgraph==0.2.60
langchain==0.3.13
langchain-openai==0.2.14
langchain-community==0.3.13
langchain-core==0.3.28
httpx==0.28.1
sqlalchemy==2.0.36
aiosqlite==0.20.0
chromadb==0.5.23
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
python-multipart==0.0.20
pyyaml==6.0.2
greenlet==3.1.1
websockets==14.1
sse-starlette==2.2.1
